PlayerWeke12
Fixed the mouse lock in different state:

At the beginning menu state, the mouse is movable and visible, so people can interact with the buttons.

After click "play" button, enter into the gameWorld, then the mouse will be locked at the center of the screen, and invisible. Instead, a crosshair will be seen at the center of the screen.